requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
attrs==25.3.0
Authlib==1.6.4
backoff==2.2.1
browser-use==0.7.10
bubus==1.5.6
cachetools==6.2.0
cdp-use==1.4.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.2
Cython==3.1.4
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.117.0
frozenlist==1.7.0
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.39.1
googleapis-common-protos==1.70.0
greenlet==3.2.4
groq==0.32.0
h11==0.16.0
html2text==2025.4.15
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mcp==1.15.0
multidict==6.6.4
oauthlib==3.3.1
ollama==0.6.0
openai==1.109.1
pillow==11.3.0
playwright==1.55.0
portalocker==2.10.1
posthog==6.7.6
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pyee==13.0.0
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyotp==2.9.0
pyparsing==3.2.5
pypdf==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
referencing==0.36.2
reportlab==4.4.4
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
rsa==4.9.1
screeninfo==0.8.1
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.48.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uuid7==0.1.0
uvicorn==0.37.0
websockets==15.0.1
yarl==1.20.1
